+++ src/main/java/xt9/deepmoblearning/common/CommonProxy.java
@@ -102,6 +102,7 @@ public class CommonProxy implements IGuiHandler {
     public Object getClientGuiElement(int ID, EntityPlayer player, World world, int x, int y, int z) {
         return null;
     }
+
     public IPlayerTrial getClientPlayerTrialCapability() {
         return null;
     }