+++ src/main/java/xt9/deepmoblearning/common/events/CapabilityHandler.java
@@ -3,6 +3,7 @@ package xt9.deepmoblearning.common.events;
 import net.minecraft.client.entity.EntityPlayerSP;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
@@ -14,7 +15,9 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 import xt9.deepmoblearning.DeepConstants;
 import xt9.deepmoblearning.client.gui.TrialOverlay;
 import xt9.deepmoblearning.common.capabilities.IPlayerTrial;
+import xt9.deepmoblearning.common.capabilities.PlayerTrial;
 import xt9.deepmoblearning.common.capabilities.PlayerTrialProvider;
+import xt9.deepmoblearning.common.util.PlayerHelper;
 
 /**
  * Created by xt9 on 2018-04-07.
@@ -37,21 +40,35 @@ public class CapabilityHandler {
     public static void onPlayerClone(PlayerEvent.Clone event) {
         EntityPlayer player = event.getEntityPlayer();
 
-        IPlayerTrial cap = player.getCapability(PlayerTrialProvider.PLAYER_TRIAL_CAP, null);
+        if(!player.world.isRemote) {
+            IPlayerTrial cap = player.getCapability(PlayerTrialProvider.PLAYER_TRIAL_CAP, null);
 
-        IPlayerTrial oldCap = event.getOriginal().getCapability(PlayerTrialProvider.PLAYER_TRIAL_CAP, null);
+            IPlayerTrial oldCap = event.getOriginal().getCapability(PlayerTrialProvider.PLAYER_TRIAL_CAP, null);
 
-        cap.setDefeated(oldCap.getDefated());
-        cap.setCurrentWave(oldCap.getCurrentWave());
-        cap.setLastWave(oldCap.getLastWave());
-        cap.setWaveMobTotal(oldCap.getWaveMobTotal());
-        cap.setTilePos(oldCap.getTilePos());
+            cap.setDefeated(oldCap.getDefated());
+            cap.setCurrentWave(oldCap.getCurrentWave());
+            cap.setLastWave(oldCap.getLastWave());
+            cap.setWaveMobTotal(oldCap.getWaveMobTotal());
+            cap.setTilePos(oldCap.getTilePos());
+        }
     }
 
 
     @SubscribeEvent
+    @SideOnly(Side.SERVER)
+    public static void onPlayerJoinedWorldServer(EntityJoinWorldEvent event) {
+        /* Re-init the player capability when player joins world */
+        if(event.getEntity() instanceof EntityPlayerMP) {
+            EntityPlayerMP p = (EntityPlayerMP) event.getEntity();
+            PlayerTrial cap = (PlayerTrial) PlayerHelper.getTrialCapability(p);
+            cap.sync(p);
+        }
+
+    }
+
+    @SubscribeEvent
     @SideOnly(Side.CLIENT)
-    public static void onPlayerJoinedWorld(EntityJoinWorldEvent event) {
+    public static void onPlayerJoinedWorldClient(EntityJoinWorldEvent event) {
         /* Re-init the player capability when player joins world */
         if(event.getEntity() instanceof EntityPlayerSP) {
             TrialOverlay.initPlayerCapability();