+++ src/main/java/xt9/deepmoblearning/common/util/PlayerHelper.java
@@ -5,7 +5,10 @@ import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.management.PlayerList;
 import net.minecraft.util.EnumHand;
+import net.minecraft.util.math.AxisAlignedBB;
+import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextComponentString;
+import net.minecraft.world.World;
 import net.minecraftforge.fml.common.FMLCommonHandler;
 import xt9.deepmoblearning.DeepMobLearning;
 import xt9.deepmoblearning.common.capabilities.IPlayerTrial;
@@ -13,14 +16,18 @@ import xt9.deepmoblearning.common.capabilities.PlayerTrial;
 import xt9.deepmoblearning.common.capabilities.PlayerTrialProvider;
 import xt9.deepmoblearning.common.items.ItemDeepLearner;
 import xt9.deepmoblearning.common.network.UpdateTrialOverlayMessage;
-
-import java.awt.*;
 import java.util.UUID;
 
 /**
  * Created by xt9 on 2017-06-14.
  */
 public class PlayerHelper {
+    public static java.util.List<EntityPlayerMP> getPlayersInArea(World world, BlockPos pos, int area, int yStart, int yEnd) {
+        BlockPos point1 = new BlockPos(pos.getX() - area, yStart, pos.getZ() - area);
+        BlockPos point2 = new BlockPos(pos.getX() + area, yEnd, pos.getZ() + area);
+        return world.getEntitiesWithinAABB(EntityPlayerMP.class, new AxisAlignedBB(point1, point2), p -> !p.isDead);
+    }
+
     @SuppressWarnings("ConstantConditions")
     public static IPlayerTrial getTrialCapability(EntityPlayerMP player) {
         if(PlayerTrialProvider.PLAYER_TRIAL_CAP != null) {
@@ -40,6 +47,11 @@ public class PlayerHelper {
         if(player != null) {
             player.sendMessage(component);
         }
+    }
+
+    public static EntityPlayerMP getPlayerFromUUID(UUID uuid) {
+        PlayerList list = FMLCommonHandler.instance().getMinecraftServerInstance().getPlayerList();
+        return list.getPlayerByUUID(uuid);
     }
 
     private boolean isHoldingDeepLearner;
@@ -62,11 +74,6 @@ public class PlayerHelper {
             stack = ItemStack.EMPTY;
         }
 
-    }
-
-    public static EntityPlayerMP getPlayerFromUUID(UUID uuid) {
-        PlayerList list = FMLCommonHandler.instance().getMinecraftServerInstance().getPlayerList();
-        return list.getPlayerByUUID(uuid);
     }